Renaming invite_institution_handler and moving invite_collection_handler's get method
