+++ frontend/invites/inviteService.js
@@ -87,7 +87,7 @@
 
         service.getSentInstitutionInvitations = function getSentInstitutionInvitations() {
             var deferred = $q.defer();
-            $http.get(INVITES_URI).then(function success(response) {
+            $http.get(INVITES_URI + '/institution').then(function success(response) {
                 deferred.resolve(response);
             }, function error(response) {
                 deferred.reject(response);
